splits form and attachments
